+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 `Unreleased`_
 -------------
 
+`4.3.0`_ - 2022-11-02
+---------------------
+
 Added
 ^^^^^
 * Added the ability to retry failed commands
@@ -405,7 +408,8 @@ Added
 * Initial project structure
 
 
-.. _`unreleased`: https://github.com/spapanik/yamk/compare/v4.2.0...main
+.. _`unreleased`: https://github.com/spapanik/yamk/compare/v4.3.0...main
+.. _`4.3.0`: https://github.com/spapanik/yamk/compare/v4.2.0...v4.3.0
 .. _`4.2.0`: https://github.com/spapanik/yamk/compare/v4.1.0...v4.2.0
 .. _`4.1.0`: https://github.com/spapanik/yamk/compare/v4.0.0...v4.1.0
 .. _`4.0.0`: https://github.com/spapanik/yamk/compare/v3.0.1...v4.0.0